+++ packages/carbon-graphs/tests/unit/controls/Graph/GraphPanningMultipleDatasets-spec.js
@@ -24,8 +24,16 @@ import {
 describe('Graph - Panning', () => {
   let graph = null;
   let graphContainer;
+  let consolewarn;
+
   beforeAll(() => {
     loadCustomJasmineMatcher();
+    // to supress warnings
+    consolewarn = console.warn;
+    console.warn = () => {};
+  });
+  afterAll(() => {
+    console.warn = consolewarn;
   });
   beforeEach(() => {
     graphContainer = document.createElement('div');