+++ packages/carbon-graphs/tests/unit/controls/PairedResult/PairedResultPanningMultipleDatasets-spec.js
@@ -25,8 +25,16 @@ import errors from '../../../../src/js/helpers/errors';
 describe('PairedResult', () => {
   let graphDefault = null;
   let pairedResultGraphContainer;
+  let consolewarn;
+
   beforeAll(() => {
     loadCustomJasmineMatcher();
+    // to supress warnings
+    consolewarn = console.warn;
+    console.warn = () => {};
+  });
+  afterAll(() => {
+    console.warn = consolewarn;
   });
   beforeEach(() => {
     pairedResultGraphContainer = document.createElement('div');